+++ SOPT-30th-Assignment/Instagram/Instagram/Constants/Const.swift
@@ -1,0 +1,12 @@
+//
+//  Const.swift
+//  Instagram
+//
+//  Created by User on 2022/04/12.
+//
+
+import Foundation
+
+struct Const {
+    
+}
