+++ SOPT-30th-Assignment/Instagram/Instagram/Constants/Image.swift
@@ -1,0 +1,18 @@
+//
+//  Image.swift
+//  Instagram
+//
+//  Created by User on 2022/04/12.
+//
+
+import UIKit
+
+extension Const {
+        
+    struct Image {
+        // 텍스트 필드 아이콘
+        static let password_shown = UIImage(named: "password_shown_eye_icon")
+        static let password_hidden = UIImage(named: "password_hidden_eye_icon")
+    }
+    
+}
