+++ SOPT-30th-Assignment/Instagram/Instagram/Constants/Storyboard.swift
@@ -1,0 +1,18 @@
+//
+//  Storyboard.swift
+//  Instagram
+//
+//  Created by User on 2022/04/12.
+//
+
+import Foundation
+
+extension Const {
+    
+    struct Storyboard {
+        //Auth
+        
+        static let LoginConfirm = "LoginConfirm"
+    }
+    
+}
